Get install path with direnv/readlink instead of from the environment
Using an environment variable does not work without ensuring that it is
set with direnv or some other method if direnv is not available,
modifying a sudoers file to have env_keep with INSTALL_PATH, and also making
every sudo call to these scripts has --preserve-env="INSTALL_PATH".
Just use a combination of direnv and readlink to get the path to the
scripts, which is also the install path since they reside in the base
directory of this repository. Even though readlink is not completely
portable, it exists on all the systems that I use.